- I was expecting the team to give me a sink pixel reference file which was formatted in the normal instrument way and was already trimmed. But they gave me one that was in the untrimmed format. This meant changing how and where the SINK pixel mask flagging was done. Instead of fixing both chips at the same time, I moved the sink detect call to just after the dodqi DQ initialization inside of wf3ccd. This meant reorganizing the sink pixel code to do flagging on one chip at a time and redoing the functions in razutil to reformat one chip at a time into and out of the RAZ orientation.
- I fixed a bug that was preventing the correct dark reference file from being used for the CTE pipeline, the checkDark code now looks to see if PCTECORR has been performed and sets the dark reference file used in wf32d to that referenced by the DARKCFILE in the header. Otherwise it uses the DARKFILE reference. This meant I could removes the darkc sets as a separate reference file. 

- Made the PCTETAB read use float for the qlev column data so that int didn't get overrun, it wasn't caught before, the table was being written with bad numbers past 32k.

- Added logic  to remove temporary files when crrej happens and cte corr is on



git-svn-id: http://svn.stsci.edu/svn/ssb/stsci_python/hstcal/trunk@40752 fe389314-cf27-0410-b35b-8c050e845b92

+++ pkg/wfc3/calwf3/wf3cte/cte.h
@@ -16,7 +16,7 @@ typedef struct {
     float thresh; /*over subtraction threshold*/        
     int cte_traps; /*number of valid TRAPS in file for reallocation*/
     int wcol_data[TRAPS]; /*trap number, insync with number of traps*/
-    int qlevq_data[TRAPS];/*charge packet size in electrons*/
+    float qlevq_data[TRAPS];/*charge packet size in electrons*/
     float dpdew_data[TRAPS];/*trap size in electrons*/  
     int   iz_data[RAZ_COLS]; /*column number in raz format*/
     double scale512[RAZ_COLS]; /*scaling appropriate at row 512 */
